Implement EmailVerificationFragment class.
This change addresses the need by creating the EmailVerification class.
The class has its own ViewModel that will keep track of the response
from server. The class implements a button listener so when it is
clicked it tells the ViewModel to hit the endpoint to get verification
code. When done the fragment will check the endpoint response with the
value in the EditText field. If successful then navigate to sign in
page.