Update theme when finishing settings activity.
Why: When settings activity finishes then the parent activity needs to
update its own color theme.

This change addresses the need by updating the call to
startActivityForResult. This will allow results to be sent back to the
parent activity and allow for communication. In the onActivityResult the
parent activity will recreate the activity, which will call the
onCreate() and then using the SharedPreferences update the theme.