Disable user sign out in the Auth activity.
Why: User should not be able to sign out at all in the sign in and register fragments.

This change addresses the need by removing the check to see if the option is selected. Nothing will be done now if the user selects the sign out option in the toolbar.